working solution

+++ src/main/java/com/lastminute/salestaxes/repository/ProductRepository.java
@@ -4,10 +4,7 @@ import com.lastminute.salestaxes.model.Product;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 
 @Repository
-public interface ProductRepository extends CrudRepository<Product, BigInteger> {
-
+public interface ProductRepository extends CrudRepository<Product, Integer> {
 }
